Moved NonLockGuarded and NonLockGuardedType to package "com.github.cloudyrock.mongock".
