+++ mongock-api/src/main/java/com/github/cloudyrock/mongock/NonLockGuardedType.java
@@ -1,4 +1,4 @@
-package io.changock.migration.api.annotations;
+package com.github.cloudyrock.mongock;
 
 public enum NonLockGuardedType {
   /**
