+++ mongock-decorator-util/src/main/java/com/github/cloudyrock/mongock/test/util/decorator/DecoratorValidator.java
@@ -5,8 +5,8 @@ import com.github.cloudyrock.mongock.driver.api.lock.guard.decorator.DecoratorBa
 import com.github.cloudyrock.mongock.driver.api.lock.guard.invoker.LockGuardInvoker;
 import com.github.cloudyrock.mongock.driver.api.lock.guard.invoker.LockGuardInvokerImpl;
 import com.github.cloudyrock.mongock.internal.DecoratorDiverted;
-import io.changock.migration.api.annotations.NonLockGuarded;
-import io.changock.migration.api.annotations.NonLockGuardedType;
+import com.github.cloudyrock.mongock.NonLockGuarded;
+import com.github.cloudyrock.mongock.NonLockGuardedType;
 import org.mockito.Mockito;
 import org.mockito.invocation.Invocation;
