+++ mongock-driver/mongock-driver-api/src/main/java/com/github/cloudyrock/mongock/driver/api/lock/guard/proxy/LockGuardProxy.java
@@ -1,8 +1,8 @@
 package com.github.cloudyrock.mongock.driver.api.lock.guard.proxy;
 
 import com.github.cloudyrock.mongock.driver.api.lock.LockManager;
-import io.changock.migration.api.annotations.NonLockGuarded;
-import io.changock.migration.api.annotations.NonLockGuardedType;
+import com.github.cloudyrock.mongock.NonLockGuarded;
+import com.github.cloudyrock.mongock.NonLockGuardedType;
 
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
