+++ mongock-driver/mongock-driver-api/src/test/java/com/github/cloudyrock/mongock/driver/api/lock/guard/proxy/util/InterfaceTypeImplNonLockGuarded.java
@@ -1,6 +1,6 @@
 package com.github.cloudyrock.mongock.driver.api.lock.guard.proxy.util;
 
-import io.changock.migration.api.annotations.NonLockGuarded;
+import com.github.cloudyrock.mongock.NonLockGuarded;
 
 @NonLockGuarded
 public class InterfaceTypeImplNonLockGuarded implements InterfaceType{
