+++ mongock-runner/mongock-runner-core/src/test/java/com/github/cloudyrock/mongock/runner/core/util/InterfaceDependencyImplNoLockGarded.java
@@ -1,6 +1,6 @@
 package com.github.cloudyrock.mongock.runner.core.util;
 
-import io.changock.migration.api.annotations.NonLockGuarded;
+import com.github.cloudyrock.mongock.NonLockGuarded;
 
 @NonLockGuarded
 public class InterfaceDependencyImplNoLockGarded implements InterfaceDependency {
