+++ mongock-runner/spring/springboot-v2.4/mongock-springboot-v2.4/src/test/java/com/github/cloudyrock/springboot/v2_2/util/InterfaceDependencyImplNoLockGarded.java
@@ -1,6 +1,6 @@
 package com.github.cloudyrock.springboot.v2_2.util;
 
-import io.changock.migration.api.annotations.NonLockGuarded;
+import com.github.cloudyrock.mongock.NonLockGuarded;
 
 @NonLockGuarded
 public class InterfaceDependencyImplNoLockGarded implements InterfaceDependency {
